[ANDROSDK-1296] Reuse hasLimitByProgram
